polish: final-presentation sweep — bibliography, typography, voice, headers, zero-warning build
- docs/REFERENCES.md: add 9 primary sources cited in code but absent
  from the bibliography (Schulten–Gordon 1975, Luscombe–Luban 1998,
  NIST DLMF §18.9.1 + §34.2, Elser 1989, Lecheminant 1997, Waldtmann
  1998, Yan–Huse–White 2011, Läuchli–Sudan–Moessner 2019, Kitaev–
  Preskill 2006, Levin–Wen 2006, Jiang–Wang–Balents 2012). Two new
  subject sections: "Frustrated magnetism" and "Topological
  entanglement entropy".

- typography: unify every author-pair spelling to en-dash (U+2013)
  across README, docs/*.md, docs/tutorials/*.md, CHANGELOG — 11
  name-pair conventions normalised.

- voice: remove first-person-plural phrasing and relative-time words
  ("now", "today") from API.md, DESIGN.md, METHODS.md,
  PHYSICS_RESULTS.md, README.md. Semantic content unchanged.

- test headers: 14 test translation units gained short coverage
  summaries at the top (types, CG, wigner_d, so3, su2, spherical
  harmonics, solid harmonics, multiset, tensor product, equivariant
  layers, nequip, parity, quadrature, radial, recoupling, time
  reversal). src/error.c gained a module header documenting the
  thread-local last-error channel.

- include/irrep/spin_project.h: stray backslash in Doxygen LaTeX
  formula (\chi_J^\*) was triggering a LaTeX error during formula
  pre-rendering. make docs now exits clean with zero warnings.

- zero-warning build: fixed -Wsign-compare in
  src/tensor_product.c:286 and removed unused expect_arrow_
  helper in src/nequip_spec.c. Release build now emits no compiler
  warnings with -Wall -Wextra -Wpedantic.

Verification:
  - make all             → 30 suites, 0 failed
  - make asan / ubsan    → 30 suites, 0 failed
  - make docs            → 0 warnings
  - make fuzz-run        → 7 × 1M iterations, 0 crashes
  - make coverage        → 91.51% line / 94.20% function (baseline)
  - make release         → ABI hash unchanged (385ef2d1…)

+++ src/tensor_product.c
@@ -283,7 +283,7 @@ tp_descriptor_t *irrep_tp_build(const irrep_multiset_t *a,
          * loops. Visit (ima, imb, imc) in the same order apply_core_
          * does so iteration is memory-sequential. */
         int n_nz = 0;
-        for (int i = 0; i < nW; ++i) if (p->cg[i] != 0.0) ++n_nz;
+        for (size_t i = 0; i < nW; ++i) if (p->cg[i] != 0.0) ++n_nz;
         p->nz   = (n_nz > 0) ? malloc((size_t)n_nz * sizeof *p->nz) : NULL;
         p->n_nz = n_nz;
         if (n_nz > 0 && !p->nz) goto fail;
